review the lighting
review the lighting

+++ 12.AirHockeyTouch/app/src/main/res/raw/texture_fragment_shader.glsl
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 
 //纹理数据
 uniform sampler2D u_TextureUnit;
@@ -8,4 +8,6 @@ varying vec2 v_TextureCoordinates;
 void main() {
 
     gl_FragColor = texture2D(u_TextureUnit,v_TextureCoordinates);
+    //gl_FragColor = vec4(0.,1.,0.,1.);
+
 }
